+++ plugins/aves_video/lib/src/slow_motion.dart
@@ -1,3 +1,5 @@
+import 'dart:ui';
+
 import 'package:aves_utils/aves_utils.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
@@ -17,8 +19,6 @@ class SlowMotionRange {
     }
     return SlowMotionRange(start: start, end: end);
   }
-
-  bool inRange(double progress) => start < progress && progress < end;
 }
 
 mixin SlowMotionMixin on Disposer {
@@ -28,6 +28,7 @@ mixin SlowMotionMixin on Disposer {
   bool get isSlowMotion => slowMotionFactor != 1;
 
   static const double fallbackPlaybackFps = 30;
+  static const int _approachDurationMillis = 300;
 
   @override
   void dispose() {
@@ -40,4 +41,31 @@ mixin SlowMotionMixin on Disposer {
   void setSlowMotionEnd(double end) => _setSlowMotion(SlowMotionRange(start: slowMotionRangeNotifier.value.start, end: end));
 
   void _setSlowMotion(SlowMotionRange v) => slowMotionRangeNotifier.value = v.sanitize();
+
+  double getSlowMotionTargetSpeed({required int currentPosition, required int duration}) {
+    double targetSpeed = 1.0;
+
+    if (duration == 0) return targetSpeed;
+
+    final range = slowMotionRangeNotifier.value;
+    var startPosition = range.start * duration;
+    var endPosition = range.end * duration;
+
+    // no approach when range is on video edges
+    if (startPosition == 0) {
+      startPosition -= _approachDurationMillis;
+    }
+    if (endPosition == duration) {
+      endPosition += _approachDurationMillis;
+    }
+
+    final startDelta = currentPosition - startPosition;
+    final endDelta = currentPosition - endPosition;
+
+    double t = startDelta.abs() < endDelta.abs() ? startDelta : -endDelta;
+    t = ((t + _approachDurationMillis) / (2 * _approachDurationMillis)).clamp(0, 1);
+
+    targetSpeed = lerpDouble(1, 1 / slowMotionFactor, roundToPrecision(t, decimals: 1)) ?? targetSpeed;
+    return targetSpeed;
+  }
 }