+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Cataloguing: detect/filter slow motion videos
+- Video: slow motion range control
 - Map: show paths
 - Settings: automatic export option